requirements.txt
factorio_rcon_py==1.2.1
requests==2.26.0
irc==19.0.1
